Add support for draw call capturing.
This commit introduces a DrawableSurface interface that describes all of
the rendering functionality necessary to render a MathExpressionSpan. It
can be used by clients to "pre-render" the LaTeX in cases where it needs
to know information ahead of actual rendering (such as drawable bounds
to estimate the necessary canvas size for rendering).